requirements.txt
Django == 4.0.4
python-telegram-bot == 13.12